+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # open890 Changelog
 
+## Unreleased
+
 ## 0.5.1 - 2021-11-22
 * Fixed non-functional macro buttons