+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # open890 Changelog
 
 ## Unreleased
+* Added SPLIT
+* Added collapsible side panel for buttons. Rearranged UI considerably.
 * Added SQL control
 * Added NR, NB, BC, NCH control buttons.
 * Added the ability to pop out separate bandscope, audioscope, and meter displays.