+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # open890 Changelog
 
-## Unreleased
+## 0.6.0 - 2021-12-05
 * Added SPLIT button, TX frequency display on spectrum, offscreen indicator
 * Rearranged UI considerably. Added collapsible side panel for buttons, arranged buttons into tabs
 * [Dev] Removed node, npm, webpack and node-sass as development dependencies in favor of esbuild and dart-sass.